+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 - `e` in keymap reads an SQL query from the minibuffer and displays the output in a temporary buffer.
 
+- In table buffers, include a button that dumps the table as CSV to an Emacs buffer.
+
 - Faces `pgmacs-table-header` and `pgmacs-table-data` are used to display the header and the rows of
   database tables.